ビルドエラー対応１

+++ DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/ChoshuHohoIdentifier.java
@@ -19,7 +19,7 @@ public class ChoshuHohoIdentifier implements Serializable {
 
     private final FlexibleYear 賦課年度;
     private final HihokenshaNo 被保険者番号;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -30,7 +30,7 @@ public class ChoshuHohoIdentifier implements Serializable {
      */
     public ChoshuHohoIdentifier(FlexibleYear 賦課年度,
             HihokenshaNo 被保険者番号,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.賦課年度 = 賦課年度;
         this.被保険者番号 = 被保険者番号;
         this.履歴番号 = 履歴番号;
