+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/ShotokuIdentifier.java
@@ -19,7 +19,7 @@ public class ShotokuIdentifier implements Serializable {
 
     private final FlexibleYear 所得年度;
     private final ShikibetsuCode 識別コード;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -30,7 +30,7 @@ public class ShotokuIdentifier implements Serializable {
      */
     public ShotokuIdentifier(FlexibleYear 所得年度,
             ShikibetsuCode 識別コード,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.所得年度 = 所得年度;
         this.識別コード = 識別コード;
         this.履歴番号 = 履歴番号;
@@ -59,7 +59,7 @@ public class ShotokuIdentifier implements Serializable {
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return 履歴番号;
     }
 }
